UUFV-365 Refactored existingApplications
