+++ libs/application/templates/parental-leave/src/fields/index.ts
@@ -19,7 +19,7 @@ export { default as InReviewSteps } from './InReviewSteps'
 export { default as Conclusion } from './PeriodsSectionImage/ConclusionImageScreen'
 export { default as DraftRequiresAction } from './DraftRequiresAction'
 export { default as PeriodsSectionImage } from './PeriodsSectionImage/PeriodsSectionImage'
-export { default as ChildSelector } from './ChildSelector'
+export { default as ExistingApplications } from './ExistingApplications'
 export { default as EditsInReviewSteps } from './InReviewSteps/EditsInReviewSteps'
 export { default as EditsRequireAction } from './EditsRequireAction'
 export { default as EmployerApprovalExtraInformation } from './EmployerApprovalExtraInformation'
