+++ libs/application/templates/parental-leave/src/forms/Prerequisites.ts
@@ -1,11 +1,14 @@
 import {
   buildAlertMessageField,
+  buildCustomField,
   buildDataProviderItem,
   buildDataProviderPermissionItem,
   buildDateField,
   buildDescriptionField,
   buildExternalDataProvider,
   buildForm,
+  buildLinkField,
+  buildMessageWithLinkButtonField,
   buildMultiField,
   buildRadioField,
   buildSection,
@@ -630,6 +633,12 @@ export const PrerequisitesForm: Form = buildForm({
 
                     return hasMultipleBirths === YES && selectedChild
                   },
+                }),
+                buildCustomField({
+                  id: 'exisitingApplications',
+                  title:
+                    parentalLeaveFormMessages.selectChild.activeApplication,
+                  component: 'ExistingApplications',
                 }),
                 buildSubmitField({
                   id: 'toDraft',
